chore(layer-list-comtrols) clean up
