+++ src/components/LayerListControls/useSelected.js
@@ -1,15 +1,7 @@
-/** @typedef {String[]} SelectedIds */
-/** @typedef {{ value: SelectedIds }} SelectedIdsRef */
 import { ref } from '@vue/composition-api'
 
-/**
- * @type {SelectedIds}
- */
 const selectedIds = ref([])
 
-/**
- * @param {SelectedIds} newList
- */
 function setSelectedIds(newList) {
   selectedIds.value = newList
 }
