+++ src/lib/add-index.js
@@ -1,22 +1,8 @@
-/** @typedef {import('./_types').Layer} */
-/** @typedef {import('./_types').LayerGroup} */
-/** @typedef {Layer & { index: Number }} LayerWithIndex */
-/** @typedef {LayerGroup & { children: LayerGroup[] | LayerWithIndex[] }} LayerGroupWithIndex */
-/** @typedef {{ value: Number }} LayerIndex */
-/**
- * @typedef Index
- * @property {() => Number} getValue
- * @property {() => Number} increment
- */
-
 /**
  * The createIndex function creates a closure around the currentIndex and
  * provides methods to increase and read the value of the currentIndex value.
  * We do this because we need to keep a reference of the current index in the
  * recursive function.
- *
- * @param {Number} value initial value
- * @returns {Index}
  */
 function createIndex(value) {
   let currentIndex = value
@@ -30,9 +16,6 @@ function createIndex(value) {
 /**
  * [Recursive]
  * Reverses all items in the array as well as the array of the `children` property
- *
- * @param {LayerGroup[] | LayerGroupWithIndex[]} list
- * @returns {LayerGroup[] | LayerGroupWithIndex[]}
  */
 function reverse(list) {
   return list
@@ -47,10 +30,6 @@ function reverse(list) {
 /**
  * [Recursive]
  * Adds indexes to all layers in a LayerGroup
- *
- * @param {LayerGroup[]} list
- * @param {Index} index
- * @returns {LayerGroupWithIndex[]}
  */
 function addIndexToLayers(list, index) {
   return list
@@ -70,20 +49,8 @@ function addIndexToLayers(list, index) {
     })
 }
 
-/**
- *
- * @param {LayerGroup[]} source
- * @returns {LayerGroupWithIndex[]}
- */
 function addIndex(source) {
-  /**
-   * @type {Index}
-   */
   let index = createIndex(-1)
-
-  /**
-   * @type {LayerGroup[]}
-   */
   const _source = JSON.parse(JSON.stringify(source))
   const reversed = reverse(_source)
   const withIndex = addIndexToLayers(reversed, index)
@@ -92,4 +59,4 @@ function addIndex(source) {
   return normalised
 }
 
-  export default addIndex
+export default addIndex
