+++ src/views/Layers.vue
@@ -61,13 +61,3 @@
     },
   }
 </script>
-
-<style lang="scss">
-  @import '~/components/AppCore/mixins.scss';
-
-  .layers .sortable-handle {
-    @include truncate;
-
-    width: 100%;
-  }
-</style>
